Explicit casts for HTTP status code

+++ lib/proto_s3/proto_s3_client.c
@@ -499,7 +499,7 @@ callback_head(void * cookie, uint8_t * buf, size_t buflen)
 
 failed:
 	/* Invoke the upstream callback. */
-	rc = (C->callback)(C->cookie, status, lens);
+	rc = (C->callback)(C->cookie, (int)status, lens);
 
 	/* Free the cookie. */
 	free(C);
